requirements.txt
Flask==3.0.0
pymongo==4.6.0
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
bcrypt==4.0.1
pandas==2.1.3
numpy==1.26.2
elasticsearch==8.11.0
beautifulsoup4==4.12.2
lxml==4.9.3
urllib3==2.1.0
pytest==7.4.3
pypdf==3.17.1
python-docx==1.1.0
